Cleaned up device file
